requirements.txt
aiosqlite==0.21.0
fastapi==0.116.1
pydantic-settings==2.10.1
SQLAlchemy==2.0.43
sqlmodel==0.0.24
uvicorn==0.35.0
bcrypt==4.3.0
python-jose==3.5.0
email-validator==2.3.0
itsdangerous==2.2.0
redis==6.4.0
fastapi-mail==1.5.0
slowapi==0.1.9
asyncpg==0.30.0
colorlog==6.9.0